requirements.txt
absl_py==1.3.0
numpy==1.23.5
pandas==1.5.2
streamlit==1.16.0
tensorflow==2.12.0
tensorflow_text==2.12.1
tqdm==4.64.1
